pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.3",
    "langchain-classic>=1.0.0",
    "langchain-groq>=1.0.0",
    "langchain-tavily>=0.2.12",
    "python-dotenv>=1.2.1",
]
